Fix structure
