+++ src/platformer/level_creator/constants.py
@@ -1,13 +1,13 @@
 # Declare some constants and variables
 
-WIDTH, HEIGHT = (800, 600)
+WIDTH, HEIGHT = (600, 400)
 FPS = 60
-BLACK      = "#000000"
-DARKGRAY   = "#404040"
-GRAY       = "#808080"
-LIGHTGRAY  = "#d3d3d3"
-WHITE      = "#FFFFFF"
-ORANGE     = "#FF6600"
-RED        = "#FF1F00"
-PURPLE     = "#800080"
+BLACK = "#000000"
+DARKGRAY = "#404040"
+GRAY = "#808080"
+LIGHTGRAY = "#d3d3d3"
+WHITE = "#FFFFFF"
+ORANGE = "#FF6600"
+RED = "#FF1F00"
+PURPLE = "#800080"
 DARKPURPLE = "#301934"
