+++ src/platformer/physics.py
@@ -1,15 +1,10 @@
 # Physics settings
 
-import pygame as pg
-from constants import *
+from constants import FPS
 
-# Physics class
-class Physics:
-
-    clock = pg.time.Clock()
-    dt = 1 / FPS
-    gravity = 1800
-    jumpHeight = -750
-    friction = 1800
-    xAcc = 3600
-    maxXVel = 480
+dt = 1 / FPS
+gravity = 1800
+jumpHeight = -750
+friction = 1800
+xAcc = 3600
+maxXVel = 480
